+++ bot/migrations/2019-04-01-142743_command_count/up.sql
@@ -1,0 +1,3 @@
+ALTER TABLE commands ADD COLUMN count INTEGER DEFAULT 0;
+INSERT INTO commands (channel, name, count, text) SELECT channel, name, count, text FROM counters;
+DROP TABLE counters;